Move pure utilities from `sort-by` to utilities package
