+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/OMC_AccessKeyDrawer.tsx
@@ -1,3 +1,4 @@
+import { useAccountSettings } from '@linode/queries';
 import {
   ActionsPanel,
   CircleProgress,
@@ -5,6 +6,7 @@ import {
   TextField,
   Typography,
 } from '@linode/ui';
+import { sortByString } from '@linode/utilities';
 import {
   createObjectStorageKeysSchema,
   updateObjectStorageKeysSchema,
@@ -15,9 +17,7 @@ import React, { useEffect, useState } from 'react';
 import { Drawer } from 'src/components/Drawer';
 import { Link } from 'src/components/Link';
 import { useObjectStorageRegions } from 'src/features/ObjectStorage/hooks/useObjectStorageRegions';
-import { useAccountSettings } from '@linode/queries';
 import { useObjectStorageBuckets } from 'src/queries/object-storage/queries';
-import { sortByString } from 'src/utilities/sort-by';
 
 import { EnableObjectStorageModal } from '../EnableObjectStorageModal';
 import { confirmObjectStorage } from '../utilities';
