+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/utils.ts
@@ -1,6 +1,4 @@
-import { areArraysEqual } from '@linode/utilities';
-
-import { sortByString } from 'src/utilities/sort-by';
+import { areArraysEqual, sortByString } from '@linode/utilities';
 
 import type { DisplayedAccessKeyScope, FormState } from './OMC_AccessKeyDrawer';
 import type { ObjectStorageKey } from '@linode/api-v4/lib/object-storage';
